requirements.txt
PySocks~=1.7.1
Telethon~=1.37.0